[inspect][validator] Pedantic 0 regions
Make sure that all parts of a block that should be 0, are 0.

Note: A decision is pending on whether FREE blocks should have 0 payload.
There's code in the Rust library and in the Validator to support and test
zeroing blocks upon Free - currently commented out and marked with
TODO(fxb/39975). (This code works and was tested, but it would take time at
runtime, and it isn't in Dart yet.)

Test: README.md
CF-911

Change-Id: I0c5d152057eeb1dacaf7b6b1a1da9f8e86e7c0b5