[mediaplayer] don't play/pause AudioRenderer without buffer
AudioRenderer now requires a payload buffer before Play, PlayNoReply,
Pause or PauseNoReply are called. This CL changes mediaplayer's
wrapper for the renderer to ensure this sequencing.

An error handler has also been added to the wrapper so the player
panics when the AudioRenderer disconnects. We may have better ways
to handle this in the future, but for now, this kills a zombie
player and prevents it from spewing fidl errors.

TEST: The fake audio renderer has been updated to have the same
      new restrictions as the real audio renderer.
Change-Id: Ia6f14429371f1814976f183679e50490ce9624f4