[escher] Precompile shaders and save to disk as SPIRV
Creates a program that can be run on the command line to precompile
all shaders used by Escher to SpirV and save the contents to disk
alongside the source shaders.

This is part one of a larger goal of being able to create builds of
Scenic that do not contain any runtime shader compiling functionality,
as the libraries required to do so are quite large and bloat scenic's
binary size.

This CL only handles the precompiling and saving to disk of the shaders,
it does not change any functionality to Scenic/Escher or makes use of
the precompiled shaders in any way, yet. That will come in a future CL,
which will allow Scenic to be built with two versions of escher, one
without shader-compiling (which will use the precompiled shaders) and
one with shader-compiling which can update shaders at runtime.

How to use:

1) fx set terminal.x64 --with //garnet/packages/examples:escher --args escher_use_null_vulkan_config_on_host=false

2) fx build host_x64/paper_shader_compiler && out/default/host_x64/paper_shader_compiler

Testing - No new unit tests since no new functionality to Scenic or Escher
has been added yet.

Change-Id: I9a217c3f0fd80d70c02045449c71d486c00624eb