[blobfs] Separate IPC dispatching from Blobfs object
Simplify the "Blobfs" main filesystem object.

- Separate the "fs::ManagedVfs" class into a "Runner" object,
which is responsible for dispatching IPC requests. This additionally
simplifies destruction of the "Blobfs" object, which can now be
logically terminated with simply a destructor instead of a dedicated
"Shutdown" method.
- Create a new dispatcher for internal blobfs requests, rather than
overloading the IPC dispatcher in the Runner.
- Remove many of the "pass-through" methods from Blobfs
to the underlying allocator; callers can access these methods
through the allocator directly.
- Privatize additional methods in the Blobfs interface.

As the filesystems move towards an "export directory" model,
it is additionally expected that global filesystem IPC dispatching
can be routed from this "Runner" class to Blobfs methods which
are decoupled from FIDL.

Change-Id: I3a8eb4594b44db77dd5f5fbb71858807a300a12b