[wlan][mlme] Consolidate TimerManager
Consolidate the various timers in C++ ClientMlme into one. This allows
us to later channel timeout events to Rust through ClientMlme instead of
Station.

Since the above change means that we lose the ability to trigger timeout
on station and channel scheduler separately in unit test, add a
ClientConfig struct to ClientMlme. This allows tests to affect the
timing of timeouts and prevents complicating our auto-deauth tests.

Test:
1. fx run-test wlan_tests
2. Manually verify that scanning and connecting still work

Change-Id: I544e47aa5cffcfe28094f6ca054b38e3036d7e1b