[zircon][cobalt-client]: Reenabling test.
Test was disabled because it was showing symptoms of poor scheduling,
where due to the number of threads and sched_yield() being called,
threads would never be scheduled again as noted in the respective
FLK bug. With the new scheduler, this problem should no be visible.
Also maniscalco suggested to reenable the test once the scheduler
migrated.

The bug is marked as fixed.

FLK-329 #comment Reenabled test

TEST=CQ

Change-Id: I82d17a39060b98c72c5b5a15540c124dec174117