[fidl][rust] ServerSender
Currently RequestStream is tightly coupling reception, deserialization
and responding to inbound FIDL requests.
The convenience of this approach comes with the cost of flexibility.
E.g., it's not possible to decode an inbound message independently from
its reception. Similarly, in order to send responses a message must have
been received and decoded using the provided structs.
The RequestMessage is dedicated for decoding a message, independent from
where the message originated from.
The ServerSender allows to send events and responses to the server end
without the need to own a ControlHandle.
These functions assist us in transitioning incrementally to Rust - a
long term goal of ours. To discourage broader usage #[deprecated] was
added.

Test:
- Updated .golden files
- Added unit test for using ServerSender and manual deserialization
  in combination with regular Rust generated bindings.

Change-Id: I5476b55c68485099f42a08be4a38e9cb0f6b0546