[cm] Remove per-realm "default_runner".
Runners will soon be specified in component manifests, with no concept
of a per-realm default runner; instead, runners will be routed/resolved
like other capabilities in the system.

We still need to have a system-wide ELF runner, which will be offered to
the root realm, which may then choose to offer it to its children.

This CL removes per-realm default runners, moving the ELF runner
instance to the Model in preparation of future changes.

Behaviour is currently unchanged: all components still get implicit
access to the ELF runner.

BUG: 4761
Change-Id: I013cc602af35635a1f263837cd88f40fa765849f