[sysmem][magma] Decrease allocated protected memory size to 32MB
This CL decreases budget to 32MB, enough to run tests and example
apps. We have seen a lot of systemwide OOM lately and this CL,
along with disabling protected memory in Chrome, should help free up
some more memory for stability. This will be increased when we are
ready to shift contiguous memory portion into protected memory.

Bug: 36620
Change-Id: I1daac6e131b21f92d35c1f3800ea53d5ea8dbad3