[ermine][dart] Migrates to fuchsia.deprecatedtimezone
This change moves the use of `fuchsia.timezone` to a functionally
equivalent `fuchsia.deprecatedtimezone`.  This is in preparation
for removing that FIDL interface completely, in favor of
the functionality provided by `fuchsia.intl.*`.

This change is a simple manipulation to reduce the number of FIDL
interfaces that we need to work on.

Bug: 39928
Change-Id: I2c940f7ec2f1c7a43cc561102428535b532aaf2b