[scenic][input] Make it so that HitTester no longer propagates hit on node to parents
HitTester currently registers a hit for every parent node of a node that gets hits.
This CL removes that side effect, which also ended up revealing a bit of unnecessary code which could be removed.
- Removes the last of the hit test tag code from HitTester.
- Removes the GlobalHitTester class since it's now always on.
Updated tests to conform to new functionality and did some test cleanup.

Adds a small hack to preserve ordering on ties, to be removed when ordering guarantees are removed in next CL.

Bug: 37712, 24102, 24126

Test: fx run-test scenic_tests -t gfx_unittests -- --gtest_filter=*HitTestTest*
Change-Id: I9c36cb4798eacc0260a63e957a67b0a84de7a829