[debugger] Cleanup "regs" command parameters.
Removes the regular expression matching capability of this command. This
diesn't work with non-canonical registers and requires a lot of work to
be able to do so. x64 vector registers are especially odd since it
displays non-canonical registers only (ymm, while canonical is zmm).

Since registers can be named in expressions, this capability is mostly
redundant. Additional output is added to tell users how to print an
individual one.

The associated filtering code can also be removed.

Change the command line arguments to be one option per type rather than
have an option for the category and force the category to be named. This
is a ltitle more flexible and less verbose.

Change-Id: I999a0d42de73ecc1758c68a9f5eeb7573d267f29