[a11y] Add more fields to semantics API
Add a more fields to semantic nodes for accessibility. This allows flutter and chromium clients to send us info about what's currently onscreen.

Note that this is not meant to be exhaustive, and there will be more CLs addition additonal fields and enum values. In particular this CL does not address most of the issue around scrolling and how we determine which node is currently focused (for various notions of focus).

BUG: 18796

Testing: fx run-test a11y_tests, fx run-test a11y_lib_tests

Change-Id: Ic61f66694dd9a2283b854d6081bf26f76999fb0c