[audio_core] fix drops in audio capture
Because of a bug in our capture-side timing calculations, we were
reading the AudioInput device's ring buffer too soon. Specifically,
we do not take the device's FIFO depth into account when we were
determining how far from the device's "Write cursor" we needed to
keep our "Capture cursor". Instead of using device-supplied values,
we use a constant worst-case value of 5 ms.

Unfortunately, Astro and Sherlock both have input FIFOs of 1024
bytes (!), which is equivalent to 5.333 ms at 48kHz stereo 16-bit.
This CL simply increases that single constant to address the drops
in audio capture data ASAP, before the appropriate fix can be
created and landed. The CL also updates numerous comments in
audio_capturer_impl.cc accordingly.

In the correct fix, we will not assume anything about fence times
for our sources, but instead use the value actually reported by the
device (FIFO depth). As sources arrive and depart, we'll need to
update this number dynamically.

Testing this simple fix is a bit more complicated, for two reasons.
First, any unit-test-style validation would really only be checking
for this particular miscalculation, and would miss other potential
problems. Second, this unique characteristic of the Astro/Sherlock
boards (FIFO depth) affects the audio stream at a point _just_
beyond the ring buffer, and thus just beyond the point that we can
detect via digital loopback. To truly validate this fix, we need to
capture audio data that came from the actual physical input, and
inspect that audio stream for dropouts (these manifest as a string
of consecutive '0' sample values).

Isn't it possible that the environment in which we capture this
audio would be very quiet, resulting in a data stream that cannot be
distinguished from 'all zeroes'? No, not really. When capturing from
the built-in microphone, its analog noise floor ensures that there
will be at least 1 bit of ongoing broad-spectrum signal (these are
not professional-grade products with noise floor of > 96dB). Thus,
if we accurately capture the incoming audio stream (i.e. if we do not
have any dropouts from reading too far), a long stretch of received
0.0 values should be extremely uncommon. If however our timing were
incorrect, there COULD be sections of the buffer that we capture
before the device had written them, which would present to us as a
consecutive stretch of audio samples with value '0.0'.

In short: to validate our capture-side mix pipeline timing, we will
record an audio buffer from the live input device, then ensure that
the longest stretch of consecutive '0.0' values received does not
exceed a defined threshold.

Note that the test captures at the audio input device's native
(default) frame_rate and channel_count, to minimize any loss in
transparency from frame-rate-conversion or rechannelization.

New test binary 'audio_core_hardware_tests' runs this specific
test case, constrained to Astro and Sherlock devices

BUG: 40183
Buganizer 143163583

Test: fx run-test audio_core_hardware_tests

MULTIPLY: `[{
    "target": "//src/media/audio/audio_core/test:audio_core_hardware_tests",
    "total_runs": 100
}]`

Change-Id: Ic764d06e927569253c1ff810186ebd57ec2dfac0