[fdio-test] deflaky fdio-test
Remove the check against the elapsed time and reply on the test framework to
handle if the test takes unreasonably long. It's not ideal, but it's a
compromise to the unreliability of the timeout mechanism in zircon.

Test:
fx set *.* --with-base //garnet/tests/zircon:zircon_tests
fx shell runtests -t fdio-test -- \
--gtest_filter=TcpSocketTest.RcvTimeout

Bug: 39870

Change-Id: Ic8ef26061241fde7d830cf233c97491107551c1c