[svchost] Move services to a background thread
Previously, all of the services ran on the same thread in svchost. Now
each service provider runs on its own thread.

This CL is a defense-in-depth change because we previously encountered
problems when one of these services was blocking the process launching
service. We no longer host the process launcher in svchost, but we might
have a similar problem in the future with other services.

Test: This CL currently has no known observable change in behavior.
Change-Id: Ib8c98d04abf160a6da75866497baa113fab2fa04