[sherlock][mmc] Sleep after issuing CMD6 in case the card is still busy
According to the eMMC spec CMD13 is not guaranteed to succeed after
issuing CMD6, so add a sleep for the maximum amount of time required as
reported by the card. Also add three retries for CMD13 just in case.
This fixes an issue on Sherlock where eMMC initialization was
occasionally failing even though the card had released the busy line.

Bug: 40091
Test: Rebooted raggi's Sherlock 100 times, eMMC initialization never
      failed. Previously it had failed approximately one in six boots.
Test: runtests -t sdmmc-test
Change-Id: I111a216cc823ad032587e5f21ec981ca4697199a