[controller][camera] Destroy the graph appropriately
CameraProcessingNode class had cyclic shared references
Setting the head pointer to nullptr was not enough
to destroy the graph.

Simplified the implementation such that going from root->leaf
all pointers are unique pointers, and the other way are raw
pointers

Made changes to the constructor that the raw pointer needs to be
initialized upon constructions and cannot change.

Test: Ran existing tests and demo multiple times to ensure there is no
      memory leak observed.

Change-Id: Icf54243e318df8550a715fbcef6ae893737c46da