[fidl][c] Transforming nullable unions
First, we were missing support for transforming nullable unions in the
transformer, and second, the typeshape was calculated incorrectly.
Because xunions are 'naturally nullable', having a nullable xunion means
that the xubnion header is directly inlined into the enclosing type.
This also means the depth is not increased more than having a
non-nullable xunion.

Test: added in fidlc, added in transformer
Change-Id: Ia8d83241f724e00e8b4401d2fe18bb717c02a791