[msd-intel-gen] Remove unnecessary CommitPages
We used to commit pages here to ensure the work was done on the
connection thread, because the BTI pin was done on the device thread.
Now that BTI pin is done on the connection thread, we can remove
the unnecessary CommitPages, which can be slow on large buffers
even when the buffer is already committed.

Test:
nuc:go/magma-tps#L1

Change-Id: If32f5308f12cb190a9ddbba02a0c94a8909c1258