[modular][intl] Get other settings from setui
Modifies the implementation of fuchsia.intl.ProfileProvider to draw the
locale settings from the 'setui' service, instead of serving static
locale settings.

This enables locales to change when they are modified by the user.

Note that this change does not change the specification of
fuchsia.intl.ProfileProvider, in which the locale setting fully
determines the hour cycle.  There is a followup change that handles that
part.

Bug: 38097
Change-Id: Ia0274733548af0d5ee7d2a07564ccef6c5260693