[debugger] Fix optimized code stepping.
This bug would happen when stepping over a deeply nested set of inline
subroutine calls. When the finish controller is running and needs to
step over an inline frame, it creates a new "step over" thread
controller.

Previously it passed the original exception type and breakpoint list
when it tells this controller to Stop. But when it had just finished
prevously stepping out of a physical frame, the exception type will be
a software breakpoint and the step over controller will stop.

This patch clears the exception type in this situation since it no
longer applies.

A clarification on the usage of the parameters is added to the thread
controller header.

Clarified several other uses of stop_type and breakpoints according to
the new requirements, alghough I don't know of current problems these
are causing.

Change-Id: I415e5bddacc34ec160cc4e1918cee076c1b89bf3