[bt][a2dp] Refactor RemotePeer, use PeerId
Move the RemotePeer out of the main into peer::Peer, and handle requests
audomatically given the set of endpoints.  This brings it more in
alignment with A2DP source and prepares for lifting functionality to a
separate library.

Use the new peer to perform endpoint discovery.

Switch to using PeerIds in place of String wherever a string was
representing a PeerId.

Fix a panic which happened when endpoint discovery failed.

Test: no behavior change
Test: manual test to confirm
Fixed: 40079
Bug: 40203

Change-Id: Ie2717ca2e8ea9cfd87d6c4221206eac74935916b