[zxtest][shadowcallstack] Fix ZxTestCase.ThreadSuspend
The test only flakes on ARM because it's related to ShadowCallStack. The
test uses zx_thread_create() and zx_thread_start() directly with a small
local buffer for the stack. On ARM we require x18 to be set up, but when
using the syscalls directly that doesn't happen.

This mightn't have been caught in the transition to ShadowCallStack
because there's a race: the main thread starts, then immediately
suspends, resumes, and kills the child thread in quick succession. So
often the child thread never gets to execute any code before it's nuked,
so doesn't get a chance to dereference the invalid x18 register.

Adding a short sleep after
https://fuchsia.googlesource.com/fuchsia/+/c39af9c3b761a61918d9c9cba5551d746dc33b60/zircon/system/ulib/zx/test/zx-test.cc#494
results in the child thread always crashing on ARM.

Because this is a test of thread suspension and isn't testing anything
to do with safestack, simply mark the child thread function as
__NO_SAFESTACK to avoid this problem.

Bug: 27339
Bug: 39288
Change-Id: Id0966e6843c3f1c0276c244945b4531321a58f8d