[magma] Commit memory on client side
Currently magma_commit_buffer commits the memory when
zx_bti_pin is called in the MSD. That's efficient, but it means that the
client needs to do magma_get_error to check for an OOM situation. If
multiple threads are committing memory at the same time, the
magma_get_error may return an error for the wrong commit since it's
global state for the device.

If we do the commit on the client side, that means we can check for an
out of memory error earlier, and know for sure which commit that's
associated with. There's also the advantage that multiple commits
(which may take a long time to allocate memory) can happen at the same
time.

magma_commit_buffer is modified to have a return value, which should be
ABI-compatible with the old usage on x64 and arm64.

Test: magma_abi_conformance_tests
Change-Id: Ia6e7389fa1820451e1bbbbd4225150109d95c15e