[scripts] Handle ".debug" extension
The build_id_conv script was meant to allow free conversion between ids.txt
and .build-id formats. It really only handled on direction correctly though
because of a minor bug that wasn't caught for a long time. It wasn't reading
handeling the .debug extension expected for debug binaries in .build-id
directories. This patch fixes that.

Change-Id: I966e0a600ea6782d2ad6fb97a8b4b1944cd49205