[netemul] Prepare before starting gRPC operations.
Stubs initiate new operation calls in one thread while
another thread pumps the gRPC completion queue.  The return
value from the stub's call is stored in a unique_ptr in a
CallData object.  If the CompletionQueue surfaces an event
on the other thread before the assignment happens, the tag
that it returns will refer to a CallData object that has
not yet assigned a value to a unique_ptr, resulting in a
nullptr dereference.

This patch calls PrepareOperation before StartCall to
ensure that the pointer is initialized before the
CompletionQueue has an opportunity to interact with it.

Bug: 36427
Bug: 37093
Bug: 38594

Change-Id: I1e5017e8f7226b1e5dc35efbae42f5c2c7498370