[i18n] Allow setui to set the hour cycle
Old code did not allow setui to set the hour cycle.  New code does.

Interestingly enough, setui does not allow for the possibilitiy of
leaving the default hour cycle setting to the locale.  So, a middle way
solution: allow the property provider to have the concept of an optional
locale; but always set from setui whatever is there.

If setui ever implements the option of a "locale driven" default hour
cycle setting, this code will continue working unmodified.

Bug: 38097
Change-Id: Ieb31f44c98a193ea7a103c8652594080be1c01b1