[fx][symbolize] Add remote symbolization
This change adds support for the symbolizer to use remote debug info buckets
to find debug information as a final last resort if symbols can't be found
locally. It stores cached files in $HOME/.build-id-cache so as to not
conflict with zxdb's .build-id directory but uses the same strategy. Eventully
it would be nice to merge these two mechnisms.

Bug: 39671

Change-Id: I5bde46d324a7759be11b75f2cacc61d9cdddf922