[fidlc] Number implicit union ordinals from 1
Previously, the xunion_ordinal for unions (ie the ordinal that the
union would have if it were a xunion) would be hashed if it is
not specified. This CL changes the xunion_ordinals to be implicitly
numbered from one, so that e.g. in

union Foo {
  int32 first;
  bool second;
};

first was have a xunion_ordinal of 1 and second would have a
xunion_ordinal of 2

Test: runtests -t fidl-simple-test
Test: runtests -t fidl-test
Test: fx run-test fidl_llcpp_types_test
Change-Id: Ia1dcb8bf9f06ffc834e7cffd8ab01327de3d3be4