[benchmarks] Pass "--quiet" to zircon_benchmarks to match the old test runner
This matches peridot/tests/benchmarks/benchmarks.cc, which runs on the
peridot-perf bot.

I forgot to add this earlier when I originally wrote the invocation in
perf_test.dart.

The reason is explained in zircon/system/ulib/perftest/runner.cc: it
reduces background noise and hence variation.  If neither invocation
used "--quiet", the resulting log messages would probably produce
different variation and overhead under the peridot-perf bot (the old
case) and under the SL4F test (the new case) because the log messages
are routed differently (via system logs and UDP for the old case; via
SSH and TCP for the new case).  But with "--quiet", the two should
behave the same.

Bug: 23091
Test: on CQ: terminal-x64-release trybot;
  local testing: "fx set terminal.x64", "fx run-e2e-tests perf_test"
Change-Id: I19beaafd0ca2a83ae936b2dd8bf44f295c9bb620