[debugger] Don't assume line table is sorted.
The previous line table searching code assumed the line table always
contains sorted addresses. But the line table can consist of address
sequences which are not sorted, some of which should be considered
stripped (they start at address 0).

Tracks the contiguous sequences of line table rows in the LineTable and
uses this to find matches.

Makes the raw row query function in the LineTable protected so callers
have to go through the sequence query to get the proper ordering and
filtering. Update the find_line.cc code accordingly.

BUG=40798

Change-Id: Id860bd50e38f9a8d744068ad28fc56ceb9b7197a