[testsharder] Require label field
Makes the test spec "label" field required by testsharder. As a result,
I had to add it to the testsharder test data.

This is a follow-up to I85e0a2b2a7df9309269adcf972a8b109c3e93f5d, which
added the "label" field. I neglected to require it as part of that CL.

Bug: 37115
Change-Id: I8c25e80b5b602e6968f1deb12a9253a0f7baa405