[escher] Fix GPU allocator unittests
Fixed: 39480
Fixed: 7270

We did the following fix:

- For Images tests we set the memory properties flags correctly so that
VmaAllocator will know that this is a host image and will allocate
host-visible memory correctly.

- For memory tests, previously it failed on AEMU because AEMU has some
virtualized memory heaps of size 512MB, which is treated by VmaAllocator
as "small heaps", and VmaAllocator allocated too much memory on that
heap.

In this change we changed the escher/BUILD.gn file to set the maximum
small heap limit to 256MB to cover this case (and also solve potential
test errors issues for devices with memory < 1GB).

- For all test cases testing allocation size, as on platforms there may
extra headers / paddings, we changed the test to ensure that for each
allocation, the allocated byte size satisfies

  kMemorySize <= bytes_allocated <= kMemorySize + kMemorySizeAllowableError

TEST=escher_tests, scenic_tests
     PASS on (1) chromebook (2) fx emu (3) host.x64

Change-Id: I42a671e24fe961cdf29e817b4aecef22257e22fc