[zircon] Allow decommit of non-writable pages that can-be-writable.
Previously VMAR's op_range(DECOMMIT) would fail with ACCESS_DENIED if
applied to non-writable VMO mappings.

Some callers render a page range inaccessible before de-committing the
pages, to guard against unexpected accesses, which doesn't work if the
mapping must be writable.

Check whether the mapping could-be-writable instead of is-writable, to
allow this usage, while still guarding against decommit of pages to
which the caller would not be able to write.

Bug: 39655
Change-Id: Ib25489b4d12e7740fd57f45764b0749473aea1d6