[sl4f][benchmarking] Add initial implementation of Dart trace model
Add an initial implementation of a new Dart trace model.  In this
change, we define the structure of the model, add logic for processing
Chrome JSON trace data into the model, and add an initial test for the
model.

Reasons for adding a new Dart trace model include:
* Implementing a Dart trace model will allow us to processes traces in
  sl4f tests without having to subprocess into external trace
  processing scripts.
* Reimplementing our trace model lets us fix some mistakes we made in
  the first one, including having only one type for all events in the
  trace model, and overusing reflection in the model's tests.

In follow-up changes, we will reimplement old metrics using this trace
model instead, implement new metrics, and finally remove the old Go
trace processing scripts.

Bug: 39301
Test: ./out/default/sl4f_client_tests
Change-Id: I1e5a02f9d082297742a624f4efb3179af14cf2e2