[astro][sherlock][thermal] Fix off-by-one error in Amlogic PWM driver
The hardware subtracts one from the high and low counts, except when the
constant enable bit is set. This was causing the duty cycle to be
slightly higher (and therefore the voltage slightly lower) than
intended. Added a unit test to ensure the register values match those
in the DVFS table in the Sherlock specifications document.

Note: This change does not seem to affect Sherlock stability.

Bug: 34733
Test: runtests -t aml-thermal-s905d2g-test
Test: Paved Astro and Sherlock.
Change-Id: Iacbb4d3a8013f43592d17c427e4ac3cc8a076420