[ledger] Use CoroutineManager in BatchUpload, fix flake
BatchUpload tries to limit the number of concurrent uploads. The logic
is much easier to follow using CoroutineManager's feature to limit the
number of concurrently running coroutines.
This avoid having to manage early exits, and we can fix 39487 in the
process.
Instead of being split into errored_/error_type_, the status is stored
as an UploadStatus in status_.

Fixed: 39487
Test: all tests still run, BatchUpload.SingleCommitWithObjectsGC
Change-Id: I2aacd652f982347f176a79ae8dfc10d70a4358d4