[testsharder] Use name, not target, for multiplier
The test "name" field in tests.json now holds the test basename rather
than the GN target name, and the testsharder multiplier "target" field
corresponds to that "name" field. So it makes sense to rename the
multiplier field to "name" for consistency and clarity.

It will still accept "target" instead of "name" for now to allow for a
soft transition. I'll stop accepting "target" after a couple weeks.

Bug: 36480
Change-Id: I7e4e0b897a48dac841bc7a1e07f348fcb22c49fe