[zircon][cmdline] Make overflowing the kernel cmdline fatal
In preparation for codifying that later-arguments-win per linked bug,
make overflowing the command line's limit fatal.

The theory here is that if there's a default specified, we always want
to be certain that later additions to the command line are handled. With
silent truncation (previous behavior) it could be possible to (for
example) accidentally not turn off a debugging feature, if the "turn it
off" setting was appended, but dropped because the base command line
later got longer.

Test: KernelCmdLineTest.Overflow* updated/added
Bug: 39661
Change-Id: I8ac31d49532e27478e2e0b63cd10b1c1b3d90893