[fbl][containers] Add sized lists.
Add versions of the doubly and singly linked lists which explicitly
keep track of list size as mutation operations are performed on the
lists.  This allows O(1) access to the count of elements in the list
instead of O(n), but comes at the cost of some extra storage, and a
small amount of extra cycles to maintain the count.

If you often need to know an exact count of members of the list, these
are probably what you want to use.  If you only really care about
whether the list is empty or not, then the tradition versions are
probably better.

Fixed: 36175

Change-Id: If4f13bde8dbef026652de70bc8ce44613302f834