[debugger] Handle const/volatile pretty printing.
The pretty printer is matched against the exact original type name and
then the concrete type name (const/volatile/typedefs removed). This
handles most cases except where the pretty-printer is matching the name
of a const typedef which won't match in the first place because of the
"const", and won't match in the second place because the typedef will be
followed.

This patch strips C-V qualifiers from the type before doing a type name
comparison.

Change-Id: I449b68f40c99b94a0810928939df254c45c022dd