[identity] Refactor AccountMap
~~ Spooky ReSuRreCtiOn 0f the UnDead AccOuntMap ~~

Context: This is a CL from June that has been revived due to
a team agreement that this would be beneficial.

AccountMap

The idea is to let shared behavior live in the same place, in order
to be easier to read and test separately. The AccountMap is currently
responsible for storing accounts to disk and lazily initializing
account handler connections when requested.

In a soon future, the plan is to traitify the AccountHandlerConnection
so that we can create a fake, and then add the majority of relevant
unit tests that are currently lacking.

Debug trait is added to AccountHandlerConnection for tests.

Change-Id: If94f2779f7ffa710a5d493cc768797eb0ce50862