[libc][asan] fix __sanitizer_fill_shadow crashes
This CL fixes an issue with the `__sanitizer_fill_shadow` function,
in which if it was being called with a small memory region, and the
threshold was low enough, it would cause weird issues due to overflows.

If `__sanitizer_fill_shadow` is called with a 0 value parameter (i.e: the
memory region is being unpoisoned), the function will try to return the
memory back to the OS, by calling `_zx_vmo_op_range` with
`ZX_VMO_OP_DECOMMIT`. However it only does this with whole pages,
calling memset with the remaining bytes before and after.

If a block smaller than a page was to be filled with zeroes (this
means that the real memory being un-poisoned was smaller than PAGE_SIZE
<< ASAN_SHADOW_SHIFT), the calculation to get the pages will be off: The
`page_start` variable (where the next aligned page starts) will be set
to an address that is after the end of the block, the `page_end`
variable will be before the block started, and `page_end - page_start`
will return a negative value, causing the decommit function to return
an error.

This is an example of how it looks like:

Addr:       0xe30fe398610 Size: 0x80
page_start: 0xe30fe398610 memset_before_size: 0x9f0
page_end:   0xe30fe398000 memset_after_size: 0x690
decommit:   0xa30fe399000 decommit_size: 0xfffffffffffff000

This CL just adds a check making sure that `shadow_size` is greater
than `PAGE_SIZE`, so that `page_start` will always fall within the range,
and the same for `page_end`.

In the case in which for example, the value to fill is exactly PAGE_SIZE,
and it falls within two pages, `page_start` will be equal to `page_end`,
and the size passed to decommit will be 0, which is fine.

I added a test case that excercises this behavior, using multiple sizes
to test edge cases, as well as different memory alignments.

To compile it, run:

```
$ fx set core.x64 \
  --with-base //bundles:tools \
  --with-base garnet/tests/zircon:zircon_tests \
  --variant=asan/sanitizer-test \
  --variant=asan/zircon_tests
$ fx build
```

And you run them with:

`$ runtests -t sanitizer-test`

Change-Id: I8cddde88a757f1a3275265214b05225e915b6d3f