[docs] Emphasizes "Using GN Build"
In the old text, it was not obvious how useful "Using GN build" is.  It
is a third hyperlink on the page, looks the same as any other and the
first two are kind of ratholing on the technical details of GN and Ninja
respectively, intended for those who already know about GN and Ninja and
need a reference manual.

It's a bit different when you are starting off.  It helps to see how
things fit together, and why.

Given how rare a basic introduction to GN is on the Internet, I think it
is helpful to emphasize it a bit.

Change-Id: Ie8c9cb6795daa3f3118374c5d534480daff2fad3