[componentmgr] Fix Hub lifetime issues
Currently ComponentManager creates the Model and Hub in a future and then
blocks indefinitely.

This is not ideal because the executor destroys the future if it detects
that it can never be woken up. This causes issues for the lifetime of
Model and the Hub, since they both exist in that future.

With this CL, ComponentManager now waits for the root component to be
destroyed rather than blocking indefinitely. Currently ComponentManager
is not notified when a process has exited, so the behavior is the same as
waiting indefinitely.

Eventually, when ComponentManager is notified on process exit, this CL will
cause ComponentManager to terminate automatically when the root component
is destroyed.

Change-Id: I37a278c13d2d14acaea9702c855d68a66dbaa374