[hermetic-compute] Fix ShadowCallStack setup
The hermetic-compute library sets up the ShadowCallStack ABI
for its captive modules.  This was starting x18 at the top of
the stack instead of the base.  ARM64's ShadowCallStack grows up.

Also change hermetic-compute to size the ShadowCallStack to match
the machine and unsafe stacks, rather than always using one page.
This matches the libc semantics.

Bug: 27339
Change-Id: Icdab7213e87eeaa35545e68b6451f3d2f53ea0ab