[brcmfmac] Remove comments about efficiecy concern
The comments in brcmf_netbuf_grow_realloc() indicates two points which
could be optimized:
1) Copy using buffer instead of whole buffer.
2) Use malloc() and free() instead of realloc and memmove to avoid
second copy.

For now we just use the safest the way, which is copying the whole
buffer, and aslo assume there is enough space for realloc() to extend
the buffer, that it will not have an extra copy during realloc(),
memmove() is the only copy.

Due to the low frequency of calling this function, the efficiency
problem of it might be picked up again when it becomes a bottleneck.

Change-Id: I4b72dd66ad326902efcacb5872514840dfe50637