[fbl] confine_array_index: Optimize x86-64 version to use conditional move
fbl::confine_array_index is used to safely bounds check array indexes in
the presence of speculative-execution related information leaks. It is used
when mitigating Spectre V1-style attacks.

The initial x86-64 version was simple and used a heavyweight LFENCE
instruction as a dispatch-serializing barrier, to force earlier branches
and bounds checks to resolve before later code. This worked, but would
have been more heavyweight than required. Switch code to use a conditional
move, creating a data dependency on the outcome of a comparison.

This brings the x86-64 code in line with the aarch64 code.

Bug: 12540 Speculative Execution Mitigations.
Bug: 33667 Spectre mitigations?
Bug: 39544 x86 Spectre V1 mitigations can be less-heavyweight than LFENCE

Change-Id: I10b88b8741771fa7e9ea87673aaa573c4a85163e