[kernel][debug][crash] add a command to intentionally crash the kernel via call recursion
The previous stack based crash call overruns the stack by allocating
increasing space on it in the form of local variables. This one overruns
the stack by recursively calling itself.

Both cases currently crash the kernel, but nice to be able to manually
trigger to make sure crash detection works.

Change-Id: Ia3ccb591a9f5b690a07cd7d45be17ee721475c13