[rust-crates] Add "futures" target and migrate build to use it
This is a necessary step towards switching fuchsia from the
futures-preview-0.3.0.alpha.19 crate to futures-0.3.0. We have some
rust users outside of the fuchsia petal, so to perform a soft migration,
this patch defines GN targets for `futures` and `futures-test` that
point to their equivalent `futures-preview` crate. Once everyone has
been migrated over to using these targets, we can remove these targets
and replace them with the real futures-0.3.0 target.

Test: CI/CQ
Change-Id: Icfd1a8d1395c7c2cfbabef7e4bf446aef903c2be