[gidl][llcpp] Fix the LLCPP build
This CL fixes the LLCPP backend of GIDL so that the conformance tests
build correctly. Before, some fields were being passed by pointer
instead of value, breaking the build.

This does not include the regenerated test files. The tests now build,
but they do not all pass, since the LLCPP backend lacks support for the
v1 wire format. That will be added in a follow-up CL.

Test: fx build host-tools/gidl && fx exec $FUCHSIA_DIR/tools/fidl/gidl-conformance-suite/regen.sh && fx build garnet/public/lib/fidl/llcpp:fidl_llcpp_conformance_test
Change-Id: I90bdeefe512944409999f4e44b42e566b9b25b06