[acpi][pwrsrc] Minimal C++ migration
This does an intentionally minimal C++ conversion of the dev-pwrsrc
ACPI device. Changes not strictly required for conversion is test code
no longer #include the implementation, but rather uses a shared header
for definitions.

Motivation for this is general cleanup of the x86 board driver, which
will be easier if everything is C++.

Test: runtests -t x86-pwrsrc-test

Change-Id: I58bce383faabcbcbad6abb6b10e9f7ac39a3ef41