[zircon][cmdline] Add fuzzer for kernel/lib/cmdline
In modifying kernel/lib/cmdline/cmdline.cc's behaviour for the linked
bug, Nick and I noted the pointer bashing was a bit tricksy.

While the kernel command line isn't subject to the atrocities of
user-provided input, perhaps a fuzzer will be able to find a logic
error, so add a simple one.

UBSan is triggered in memmem by this fuzzer for 0xff<<24 due to int
promotion + overflow; cast the char to uint32_t in memmem before
shifting to avoid this. I think this would probably come up in any
fuzzer that provides its input to strncmp, so maybe we're not using
UBSan much yet?

Bug: 39661
Change-Id: Id4f02f29994769d5cc28c7eb731fadfd3c1e2a7e